+++ fabric/src/main/java/de/cech12/usefulhats/mixin/FishingRodHookedTriggerMixin.java
@@ -1,7 +1,7 @@
 package de.cech12.usefulhats.mixin;
 
 import de.cech12.usefulhats.UsefulHatsEventUtils;
-import net.minecraft.advancements.criterion.FishingRodHookedTrigger;
+import net.minecraft.advancements.triggers.FishingRodHookedTrigger;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.projectile.FishingHook;
 import net.minecraft.world.item.ItemStack;