+++ neoforge/src/main/java/de/cech12/usefulhats/client/compat/TrinketsClientCompat.java
@@ -1,5 +1,5 @@
 package de.cech12.usefulhats.client.compat;
-
+/*
 import com.mojang.blaze3d.vertex.PoseStack;
 import de.cech12.usefulhats.client.AbstractUsefulHatsRenderer;
 import de.cech12.usefulhats.platform.Services;
@@ -33,3 +33,4 @@ public class TrinketsClientCompat {
     }
 
 }
+ */