+++ neoforge/src/main/java/de/cech12/usefulhats/compat/TrinketsCompat.java
@@ -1,5 +1,5 @@
 package de.cech12.usefulhats.compat;
-
+/*
 import de.cech12.usefulhats.UsefulHatsEventUtils;
 import de.cech12.usefulhats.item.AbstractHatItem;
 import de.cech12.usefulhats.platform.Services;
@@ -58,3 +58,4 @@ public class TrinketsCompat {
     }
 
 }
+ */